[Presentation] Split files into smaller ones
