+++ lib/presentation/transaction/widgets/form_repetition_cycle_dropdown.dart
@@ -5,9 +5,7 @@ import 'package:my_expenses/domain/enums/enums.dart';
 import 'package:my_expenses/domain/models/models.dart';
 import 'package:my_expenses/domain/utils/date_utils.dart' as utils;
 import 'package:my_expenses/generated/l10n.dart';
-import 'package:my_expenses/presentation/shared/common_dropdown_button.dart';
 import 'package:my_expenses/presentation/shared/extensions/i18n_extensions.dart';
-import 'package:my_expenses/presentation/shared/utils/enum_utils.dart';
 
 const _repetitionCycles = [
   RepetitionCycleType.none,
@@ -37,51 +35,62 @@ class FormRepetitionCycleDropDown extends StatelessWidget {
   Widget build(BuildContext context) {
     final i18n = S.of(context);
     final theme = Theme.of(context);
+    final cyclesToUse = isParentTransaction
+        ? _repetitionCycles.where((c) => c != RepetitionCycleType.none).toList()
+        : _repetitionCycles.toList();
 
-    //This is to avoid loosing the isParentTransaction property and
-    //to avoid a potential bug
-    final repetitionCyclesToUse = isParentTransaction ? _repetitionCycles.where((c) => c.index != RepetitionCycleType.none.index) : _repetitionCycles;
-    return Container(
-      margin: const EdgeInsets.symmetric(vertical: 10),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: <Widget>[
-          Row(
-            children: <Widget>[
-              const Icon(Icons.repeat, size: 30),
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.only(left: 8),
-                  child: CommonDropdownButton<RepetitionCycleType>(
-                    values: EnumUtils.getTranslatedAndSortedEnum<RepetitionCycleType>(
-                      repetitionCyclesToUse,
-                      (value, _) => i18n.translateRepetitionCycleType(value),
-                    ),
-                    currentValue: repetitionCycle,
-                    hint: i18n.translateRepetitionCycleType(repetitionCycle),
-                    onChanged: (v, _) => !isChildTransaction ? _repetitionCycleChanged(v, context) : null,
-                  ),
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        ListTile(
+          contentPadding: EdgeInsets.zero,
+          enabled: !isChildTransaction,
+          leading: const Icon(Icons.repeat),
+          title: Text(i18n.recurring),
+          subtitle: Text(i18n.translateRepetitionCycleType(repetitionCycle)),
+          onTap: isChildTransaction ? null : () => _showPicker(context, cyclesToUse),
+        ),
+        if (!isChildTransaction && repetitionCycle != RepetitionCycleType.none)
+          Padding(
+            padding: const EdgeInsets.only(bottom: 5, left: 8, right: 8),
+            child: Text(
+              i18n.recurringTransactionStartsOn(
+                utils.DateUtils.formatAppDate(
+                  transactionDate,
+                  language,
+                  utils.DateUtils.monthDayAndYearFormat,
                 ),
+                i18n.translateRepetitionCycleType(repetitionCycle).toLowerCase(),
               ),
-            ],
-          ),
-          if (!isChildTransaction && repetitionCycle != RepetitionCycleType.none)
-            Padding(
-              padding: const EdgeInsets.only(bottom: 5, left: 8, right: 8),
-              child: Text(
-                i18n.recurringTransactionStartsOn(
-                  utils.DateUtils.formatAppDate(transactionDate, language, utils.DateUtils.monthDayAndYearFormat),
-                  i18n.translateRepetitionCycleType(repetitionCycle).toLowerCase(),
-                ),
-                textAlign: TextAlign.center,
-                style: theme.textTheme.bodySmall!.copyWith(color: theme.colorScheme.primary),
-              ),
+              textAlign: TextAlign.center,
+              style: theme.textTheme.bodySmall!.copyWith(color: theme.colorScheme.primary),
             ),
-        ],
-      ),
+          ),
+      ],
     );
   }
 
-  void _repetitionCycleChanged(RepetitionCycleType newValue, BuildContext context) =>
-      context.read<TransactionFormBloc>().add(TransactionFormEvent.repetitionCycleChanged(repetitionCycle: newValue));
+  void _showPicker(BuildContext context, List<RepetitionCycleType> cycles) {
+    final bloc = context.read<TransactionFormBloc>();
+    final i18n = S.of(context);
+    showModalBottomSheet(
+      context: context,
+      builder: (_) => SafeArea(
+        child: ListView(
+          shrinkWrap: true,
+          children: [
+            for (final cycle in cycles)
+              ListTile(
+                title: Text(i18n.translateRepetitionCycleType(cycle)),
+                selected: cycle == repetitionCycle,
+                onTap: () {
+                  bloc.add(TransactionFormEvent.repetitionCycleChanged(repetitionCycle: cycle));
+                  Navigator.of(context).pop();
+                },
+              ),
+          ],
+        ),
+      ),
+    );
+  }
 }
